testfix1

+++ examples/ldns-signzone.c
@@ -209,7 +209,7 @@ main(int argc, char *argv[])
 
 			break;
 		case 't':
-			nsec3_iterations_cmd = atol(optarg);
+			nsec3_iterations_cmd = (size_t) atol(optarg);
 			if (nsec3_iterations_cmd > LDNS_NSEC3_MAX_ITERATIONS) {
 			  fprintf(stderr, "Iterations count can not exceed %u, quitting\n", LDNS_NSEC3_MAX_ITERATIONS);
 			  exit(EXIT_FAILURE);
